move quill module configs into seperate hooks
